Some minor cosmetic changes. Also fixed bug.
On the cosmetic front, deleted some commented code.

In utils.py/bucketize, values from math.ceil need to be explicitly cast
to ints before being passed to range. Otherwise the summary page will hang.

+++ www/doitweb/doit/util.py
@@ -22,7 +22,7 @@ def bucketize(vals):
     if len(nums) <= 1:
         return dict()
 
-    n_buckets = math.ceil(math.sqrt(len(nums)))
+    n_buckets = int(math.ceil(math.sqrt(len(nums))))
 
     dv = stddev(nums)
     av = avg(nums)
